move and use 'views' folder
