+++ app/src/Components/views/ComponentC/ComponentC.tsx
@@ -1,4 +1,4 @@
-import { CommonFullCont } from "../../assets/common-styles/common.styles";
+import { CommonFullCont } from '../../../assets/common-styles/common.styles';
 
 const ComponentC = () => {
   return <CommonFullCont>Component C</CommonFullCont>;
